Fix C++ warnings.

+++ codecs.c
@@ -224,7 +224,7 @@ char *decode_quoted_printable(unsigned char *buf, unsigned int len, unsigned int
                   *dest++ = c;
                   break;
             }
-            if ((dest-deststart)>= limit)
+            if ((unsigned int)(dest-deststart)>= limit)
             {
             	text = NULL;
             	break;
